+++ packages/orq-rc/RELEASES.md
@@ -2119,3 +2119,13 @@ Based on:
 - [typescript v3.8.0-rc.18] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.18 - packages/orq-rc
+
+## 2025-06-12 07:22:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.561.0 (2.628.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.19] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.19 - packages/orq-rc